requirements.txt
asgiref==3.9.1
certifi==2025.10.5
cffi==2.0.0
decorator==5.2.1
defusedxml==0.7.1
Django==5.2.4
django-cors-headers==4.9.0
django-phonenumber-field==8.1.0
djangorestframework==3.16.1
drf-nested-routers==0.95.0
drf-yasg==1.21.11
et_xmlfile==2.0.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
psycopg2-binary==2.9.11
pure_eval==0.2.3
pycparser==2.23
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2025.2
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3

